requirements.txt
pandas
numpy
torch
sklearn.preprocessing
matplotlib.pyplot